[debug] Add auto crash dumps for Windows
