+++ src/common/debug.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace debug
+{
+    void init();
+} // namespace debug
