+++ src/common/debug_linux.cpp
@@ -1,0 +1,6 @@
+#include "debug.h"
+
+void debug::init()
+{
+
+}
