+++ src/common/debug_osx.cpp
@@ -1,0 +1,6 @@
+#include "debug.h"
+
+void debug::init()
+{
+
+}
